requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
django-bootstrap5==24.3
django-stdimage==6.0.2
gunicorn==23.0.0
mysql==0.0.3
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyMySQL==1.1.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
